Add text input allowing for dynamic text updating.
Added a text input field to the HTML. Through this input users can
change the animated text on screen to whatever they would like it
to say. In doing so, I changed a few variables in init() to be global
and added changeText() to be responsible for creating the text on screen.